moves Caster and Blaster into their own sub-packages; TODOs to do generic type checking in Caster
