+++ src/com/bluegosling/reflect/caster/AmbiguousMethodException.java
@@ -1,4 +1,4 @@
-package com.bluegosling.reflect;
+package com.bluegosling.reflect.caster;
 
 import java.lang.reflect.Method;
 import java.util.Collection;
