+++ src/com/bluegosling/reflect/caster/DispatchSettings.java
@@ -1,4 +1,4 @@
-package com.bluegosling.reflect;
+package com.bluegosling.reflect.caster;
 
 import java.lang.reflect.Modifier;
