+++ src/com/bluegosling/reflect/package-info.java
@@ -2,8 +2,8 @@
  * Utilities to aid in runtime type reflection code, including the use of proxies. This includes
  * numerous useful tools for working with {@linkplain com.bluegosling.reflect.Annotations annotations}
  * and generic {@linkplain com.bluegosling.reflect.Types types}. It also includes a few interesting
- * tools that rely on reflection and proxies, like a {@linkplain com.bluegosling.reflect.Caster
- * dynamic-cast proxy} or a {@linkplain com.bluegosling.reflect.Blaster broadcasting proxy}.
+ * tools that rely on reflection and proxies, like a {@linkplain com.bluegosling.reflect.caster.Caster
+ * dynamic-cast proxy} or a {@linkplain com.bluegosling.reflect.blaster.Blaster broadcasting proxy}.
  * 
  * @see com.bluegosling.apt.reflect
  */
